✨상세 페이지 하단 시리즈 캐러샐에 데이터 연동 ♻️ 시리즈 박스에 연관된 컴포넌트 리팩토링 ✨댓글, 대댓글 생성 접근 제어 ♻️ 타입 리팩토링
